refactor(Post): modify byUser method to accept user ID parameter
Update the byUser method in the Post model to accept a user ID parameter instead of using Auth::user()->id directly. This change improves flexibility and testability by decoupling the method from the authentication context. Additionally, update related views and Livewire components to pass the user ID explicitly.

+++ resources/views/livewire/main/profile/_modal-show-post.blade.php
@@ -78,9 +78,11 @@
 
 
                 <div class="py-2 justify-end flex">
-                    <x-buttons.button iconClass="ti ti-trash" color="danger" wire:click="deletePost('{{ $showPost?->id }}')" target="deletePost">
-                        Hapus Postingan
-                    </x-buttons.button>
+                @if ($showPost?->userable_id == Auth::user()->id) 
+                        <x-buttons.button iconClass="ti ti-trash" color="danger" wire:click="deletePost('{{ $showPost?->id }}')" target="deletePost">
+                            Hapus Postingan
+                        </x-buttons.button>
+                @endif
                 </div>
         @endif
     </x-ts-modal>
